requirements.txt
streamlit>=1.24.0
supabase>=2.22.2
pandas>=2.0.0
plotly>=5.13.0
python-dotenv>=1.0.0
google-generativeai>=0.2.0
stripe>=5.4.0
